requirements.txt
numpy>=1.4
biopython>=1.7
sklearn>=0
ete3>=3.1
six>=1.11
